+++ test/jest/__mock__/index.js
@@ -4,3 +4,4 @@ import './stripesCore.mock';
 import './stripesIcon.mock';
 import './reactIntl.mock';
 import './matchMedia.mock';
+import './react-virtualized-auto-sizer.mock';