+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Avoid private paths in stripes-core imports. Fixes STDTC-81.
 * Format numbers in "totalRecords" columns. STDTC-77
+* Fix tests due to changes in react-virtualized-auto-sizer. STDTC-83
 
 ## [5.4.1](https://github.com/folio-org/stripes-data-transfer-components/tree/v5.4.1) (2023-03-16)
 [Full Changelog](https://github.com/folio-org/stripes-data-transfer-components/compare/v5.4.0...v5.4.1)